+++ x64barebones/Kernel/syscalls/sys_list_process.c
@@ -1,7 +1,7 @@
 #include <multitasking.h>
 #include <syscalls.h>
 
-unsigned int sys_list_process() {
+uint64_t sys_list_process() {
 	list_process();
 	return 1;
 }
