+++ x64barebones/Kernel/syscalls/sys_register_sem.c
@@ -1,7 +1,7 @@
 #include <semaphore.h>
 #include <syscalls.h>
 
-unsigned int sys_register_sem(unsigned int sem_id, unsigned int value)
+uint64_t sys_register_sem(unsigned int sem_id, unsigned int value)
 {
 	return create_sem(sem_id, value);
 }
