+++ x64barebones/Kernel/syscalls/sys_renounce_cpu.c
@@ -2,7 +2,7 @@
 #include <syscalls.h>
 
 
-unsigned int sys_renounce_cpu(){
+uint64_t sys_renounce_cpu(){
 	forceTimerTick();
 	return 1;
 }
