+++ x64barebones/Kernel/syscalls/sys_wait_for_children.c
@@ -2,7 +2,7 @@
 #include <syscalls.h>
 
 
-unsigned int sys_wait_for_children(){
+uint64_t sys_wait_for_children(){
 	wait_for_children();
 	return 1;
 }
