+++ x64barebones/Kernel/syscalls/sys_wait_sem.c
@@ -2,6 +2,6 @@
 #include <syscalls.h>
 
 
-unsigned int sys_wait_sem(unsigned int sem_id){
+uint64_t sys_wait_sem(unsigned int sem_id){
 	return wait_sem(sem_id);
 }
